feat: refactor payload handling and add new entity interfaces
