+++ internal/payload/entity_test.go
@@ -1,4 +1,4 @@
-package scanning
+package payload
 
 import (
 	"testing"
